# zoo/dao-governance/stack/dao/api/packages/decent-offchain/src/api/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DAO API",
    "description": "API for the DAO protocol",
    "version": "0.1.0",
    "contact": {
      "name": "DAO DAO",
      "url": "https://luxdao.org"
    }
  },
  "servers": [
    {
      "url": "https://api.dao.build",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3005",
      "description": "Localhost"
    }
  ],
  "paths": {
...
    "/points/proposals": {
      "get": {
        "tags": ["points"],
        "summary": "Get proposal count for address",
        "description": "Get the number of proposals that an address has submitted",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Wallet address"
          },
          {
            "name": "passed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Whether to filter for passed proposals"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "proposalCount": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["success", "failed"]
                    }
                  },
                  "required": ["address", "proposalCount", "status"]
                }
              }
            }
          }
        }
      }
    },
    "/points/transfers": {
      "get": {
        "tags": ["points"],
        "summary": "Get transfer proposal count for address",
        "description": "Get the number of passing transfer proposals that an address has submitted",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Wallet address"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "transferCount": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["success", "failed"]
                    }
                  },
                  "required": ["address", "transferCount", "status"]
                }
              }
            }
          }
        }
      }
    },
    "/points/disperses": {
      "get": {
        "tags": ["points"],
        "summary": "Get disperse proposal count for address",
        "description": "Get the number of passing disperse proposals that an address has submitted",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Wallet address"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "disperseCount": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["success", "failed"]
                    }
                  },
                  "required": ["address", "disperseCount", "status"]
                }
              }
            }
          }
        }
      }
    },
    "/points/roles": {
      "get": {
        "tags": ["points"],
        "summary": "Get role proposal count for address",
        "description": "Get the number of passing role proposals that an address has submitted",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Wallet address"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "roleCount": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["success", "failed"]
                    }
                  },
                  "required": ["address", "roleCount", "status"]
                }
              }
            }
          }
        }
      }
    },
    "/points/streams": {
      "get": {
        "tags": ["points"],
        "summary": "Get stream proposal count for address",
        "description": "Get the number of passing stream proposals that an address has submitted",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Wallet address"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "streamCount": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["success", "failed"]
                    }
                  },
                  "required": ["address", "streamCount", "status"]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}